MARL folder and Dict for all types
