+++ cares_reinforcement_learning/algorithm/marl/__init__.py
@@ -1,1 +1,8 @@
 from .MultiMARL import MultiMARL
+from .MADDPG import MADDPG
+from .M3DDPG import M3DDPG
+from .MATD3 import MATD3
+from .MASAC import MASAC
+from .MAPPO import MAPPO
+from .IMARL import IDDPG, ITD3, ISAC, IPPO
+from .QMIX import QMIX
