+++ cares_reinforcement_learning/algorithm/policy/__init__.py
@@ -24,9 +24,3 @@ from .CrossQ import CrossQ
 from .DroQ import DroQ
 from .SDAR import SDAR
 from .TD7 import TD7
-from .MADDPG import MADDPG
-from .M3DDPG import M3DDPG
-from .MATD3 import MATD3
-from .MASAC import MASAC
-from .MAPPO import MAPPO
-from .IMARL import IDDPG, ITD3, ISAC, IPPO
