+++ cares_reinforcement_learning/algorithm/value/__init__.py
@@ -6,4 +6,3 @@ from .PERDQN import PERDQN
 from .C51 import C51
 from .QRDQN import QRDQN
 from .Rainbow import Rainbow
-from .QMIX import QMIX
